Added sensor_file
Sensor file and sensor_task will take care of all sensor data and setup of sensorpins. along with this, provides API to the rest of the program to read sensordata from a library. direct access to sensors is reserved for the sensor task and corresponding functions. added 'sensordata' keyword to uart, to print every single active data set from the sensor data library to the UART terminal. it's buggy as **** in the current build though.